+++ packages/template-retail-react-app/app/components/_app/index.jsx
@@ -53,10 +53,10 @@ import {DEFAULT_SITE_TITLE, HOME_HREF, THEME_COLOR} from '../../constants'
 
 import Seo from '../seo'
 import {resolveSiteFromUrl} from '../../utils/site-utils'
+import {BasketsProvider} from '../../commerce-hooks/context/shopper-basket-context'
 
 const DEFAULT_NAV_DEPTH = 3
 const DEFAULT_ROOT_CATEGORY = 'root'
-
 
 const App = (props) => {
     const {children, targetLocale, messages, categories: allCategories = {}} = props
@@ -172,105 +172,121 @@ const App = (props) => {
                 defaultLocale="en-US"
             >
                 <SCAPIProvider>
-                    <CategoriesProvider categories={allCategories}>
-                        <CurrencyProvider currency={currency}>
-                            <Seo>
-                                <meta name="theme-color" content={THEME_COLOR} />
-                                <meta name="apple-mobile-web-app-title" content={DEFAULT_SITE_TITLE} />
-                                <link
-                                    rel="apple-touch-icon"
-                                    href={getAssetUrl('static/img/global/apple-touch-icon.png')}
-                                />
-                                <link rel="manifest" href={getAssetUrl('static/manifest.json')} />
+                    <BasketsProvider>
+                        <CategoriesProvider categories={allCategories}>
+                            <CurrencyProvider currency={currency}>
+                                <Seo>
+                                    <meta name="theme-color" content={THEME_COLOR} />
+                                    <meta
+                                        name="apple-mobile-web-app-title"
+                                        content={DEFAULT_SITE_TITLE}
+                                    />
+                                    <link
+                                        rel="apple-touch-icon"
+                                        href={getAssetUrl('static/img/global/apple-touch-icon.png')}
+                                    />
+                                    <link
+                                        rel="manifest"
+                                        href={getAssetUrl('static/manifest.json')}
+                                    />
 
-                                {/* Urls for all localized versions of this page (including current page)
+                                    {/* Urls for all localized versions of this page (including current page)
                                 For more details on hrefLang, see https://developers.google.com/search/docs/advanced/crawling/localized-versions */}
-                                {site.l10n?.supportedLocales.map((locale) => (
+                                    {site.l10n?.supportedLocales.map((locale) => (
+                                        <link
+                                            rel="alternate"
+                                            hrefLang={locale.id.toLowerCase()}
+                                            href={`${appOrigin}${getPathWithLocale(locale.id, {
+                                                location
+                                            })}`}
+                                            key={locale.id}
+                                        />
+                                    ))}
+                                    {/* A general locale as fallback. For example: "en" if default locale is "en-GB" */}
                                     <link
                                         rel="alternate"
-                                        hrefLang={locale.id.toLowerCase()}
-                                        href={`${appOrigin}${getPathWithLocale(locale.id, {
-                                            location
-                                        })}`}
-                                        key={locale.id}
+                                        hrefLang={site.l10n.defaultLocale.slice(0, 2)}
+                                        href={`${appOrigin}${getPathWithLocale(
+                                            site.l10n.defaultLocale,
+                                            {
+                                                location
+                                            }
+                                        )}`}
                                     />
-                                ))}
-                                {/* A general locale as fallback. For example: "en" if default locale is "en-GB" */}
-                                <link
-                                    rel="alternate"
-                                    hrefLang={site.l10n.defaultLocale.slice(0, 2)}
-                                    href={`${appOrigin}${getPathWithLocale(site.l10n.defaultLocale, {
-                                        location
-                                    })}`}
-                                />
-                                {/* A wider fallback for user locales that the app does not support */}
-                                <link rel="alternate" hrefLang="x-default" href={`${appOrigin}/`} />
-                            </Seo>
+                                    {/* A wider fallback for user locales that the app does not support */}
+                                    <link
+                                        rel="alternate"
+                                        hrefLang="x-default"
+                                        href={`${appOrigin}/`}
+                                    />
+                                </Seo>
 
-                            <ScrollToTop />
+                                <ScrollToTop />
 
-                            <Box id="app" display="flex" flexDirection="column" flex={1}>
-                                <SkipNavLink zIndex="skipLink">Skip to Content</SkipNavLink>
+                                <Box id="app" display="flex" flexDirection="column" flex={1}>
+                                    <SkipNavLink zIndex="skipLink">Skip to Content</SkipNavLink>
 
-                                <Box {...styles.headerWrapper}>
-                                    {!isCheckout ? (
-                                        <Header
-                                            onMenuClick={onOpen}
-                                            onLogoClick={onLogoClick}
-                                            onMyCartClick={onCartClick}
-                                            onMyAccountClick={onAccountClick}
-                                            onWishlistClick={onWishlistClick}
+                                    <Box {...styles.headerWrapper}>
+                                        {!isCheckout ? (
+                                            <Header
+                                                onMenuClick={onOpen}
+                                                onLogoClick={onLogoClick}
+                                                onMyCartClick={onCartClick}
+                                                onMyAccountClick={onAccountClick}
+                                                onWishlistClick={onWishlistClick}
+                                            >
+                                                <HideOnDesktop>
+                                                    <DrawerMenu
+                                                        isOpen={isOpen}
+                                                        onClose={onClose}
+                                                        onLogoClick={onLogoClick}
+                                                        root={allCategories[DEFAULT_ROOT_CATEGORY]}
+                                                    />
+                                                </HideOnDesktop>
+
+                                                <HideOnMobile>
+                                                    <ListMenu
+                                                        root={allCategories[DEFAULT_ROOT_CATEGORY]}
+                                                    />
+                                                </HideOnMobile>
+                                            </Header>
+                                        ) : (
+                                            <CheckoutHeader />
+                                        )}
+                                    </Box>
+
+                                    {!isOnline && <OfflineBanner />}
+                                    <AddToCartModalProvider>
+                                        <SkipNavContent
+                                            style={{
+                                                display: 'flex',
+                                                flexDirection: 'column',
+                                                flex: 1,
+                                                outline: 0
+                                            }}
                                         >
-                                            <HideOnDesktop>
-                                                <DrawerMenu
-                                                    isOpen={isOpen}
-                                                    onClose={onClose}
-                                                    onLogoClick={onLogoClick}
-                                                    root={allCategories[DEFAULT_ROOT_CATEGORY]}
-                                                />
-                                            </HideOnDesktop>
+                                            <Box
+                                                as="main"
+                                                id="app-main"
+                                                role="main"
+                                                display="flex"
+                                                flexDirection="column"
+                                                flex="1"
+                                            >
+                                                <OfflineBoundary isOnline={false}>
+                                                    {children}
+                                                </OfflineBoundary>
+                                            </Box>
+                                        </SkipNavContent>
 
-                                            <HideOnMobile>
-                                                <ListMenu root={allCategories[DEFAULT_ROOT_CATEGORY]} />
-                                            </HideOnMobile>
-                                        </Header>
-                                    ) : (
-                                        <CheckoutHeader />
-                                    )}
+                                        {!isCheckout ? <Footer /> : <CheckoutFooter />}
+
+                                        <AuthModal {...authModal} />
+                                    </AddToCartModalProvider>
                                 </Box>
-
-                                {!isOnline && <OfflineBanner />}
-                                <AddToCartModalProvider>
-                                    <SkipNavContent
-                                        style={{
-                                            display: 'flex',
-                                            flexDirection: 'column',
-                                            flex: 1,
-                                            outline: 0
-                                        }}
-                                    >
-                                        <Box
-                                            as="main"
-                                            id="app-main"
-                                            role="main"
-                                            display="flex"
-                                            flexDirection="column"
-                                            flex="1"
-                                        >
-                                            <OfflineBoundary isOnline={false}>
-                                                {children}
-                                            </OfflineBoundary>
-                                        </Box>
-                                    </SkipNavContent>
-
-                                    {!isCheckout ? <Footer /> : <CheckoutFooter />}
-
-                                    <AuthModal {...authModal} />
-                                </AddToCartModalProvider>
-                            </Box>
-
-                        </CurrencyProvider>
-                    </CategoriesProvider>
+                            </CurrencyProvider>
+                        </CategoriesProvider>
+                    </BasketsProvider>
                 </SCAPIProvider>
             </IntlProvider>
         </Box>
