+++ packages/template-retail-react-app/app/pages/home/index.jsx
@@ -61,22 +61,21 @@ const Home = ({isLoading}) => {
         // if (res) {
         //     res.set('Cache-Control', `max-age=${MAX_CACHE_AGE}`)
         // }
-    
+
         const productSearchResult = await api.shopperSearch.productSearch({
             parameters: {
                 refine: [`cgid=${HOME_SHOP_PRODUCTS_CATEGORY_ID}`, 'htype=master'],
                 limit: HOME_SHOP_PRODUCTS_LIMIT
             }
         })
-    
+
         return productSearchResult
     }, [])
-    
+
     // NOTE: This is an example of a hook that was created using the ServerEffects API that I added
     // to the SDK. This API allows you to create a server hook context and an associated hook that you
     // can used to build your own hooks off of. You can see how what is done in the `scapi-hooks.js` file
     // which is a precursor to a larger hooks library.
-    const {product, isLoading: isUseProductLoading} = useProduct(1, [])
 
     return (
         <Box data-testid="home-page" layerStyle="page">
@@ -162,7 +161,6 @@ const Home = ({isLoading}) => {
                 </SimpleGrid>
             </Section>
 
-            {product && <div>{product.name}</div>}
             {productSearchResult && (
                 <Section
                     padding={4}
